requirements.txt
Beautifulsoup4
requests
lxml
pyTelegramBotAPI
python-dotenv
